+++ CHANGELOG.md
@@ -12,6 +12,9 @@ Since 0.19.0-beta, we use [this changelog format](https://keepachangelog.com). I
 * Data version has been increased to `7`.
 * Data in the config files can now be typed, meaning that only a certain type of data can be written.
 
+### Removed
+* `(Internal API/Writing libs)` Removed `TRUE` and `FALSE` values, please use `Number(True)` and `Number(False)` instead.
+
 ## 0.20.0-beta (2024-06-06)
 
 This version may in some cases improve performance.